feat(frontend): introduce new article components and update styles
- Added a new SVG icon for topic breadcrumbs.
- Refactored article module to utilize new components such as TitleHero and HeroImage.
- Implemented a utility function to parse post content and reorganized article rendering logic.
- Removed deprecated CSS and components to streamline the article module.
