+++ packages/frontend/src/components/image-with-fallback.tsx
@@ -7,7 +7,15 @@ export const ImageWithFallback = (
   return (
     <img
       {...props}
+      src={props.src || FALLBACK_IMG}
       onError={({ currentTarget }) => {
+        if (
+          currentTarget.src === FALLBACK_IMG ||
+          currentTarget.src.endsWith(FALLBACK_IMG)
+        ) {
+          currentTarget.onerror = null
+          return
+        }
         currentTarget.onerror = null
         currentTarget.src = FALLBACK_IMG
         currentTarget.srcset = ''
