--- packages/frontend/src/modules/article/article.css
@@ -1,388 +1,0 @@
-.skeleton {
-  width: 80vw;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 20px;
-}
-
-body.single-post {
-  --theme-color: #27b5f7;
-}
-
-body.single-post .entry-content {
-  --block-wide-max-width: 1000px;
-}
-
-@media screen and (max-width: 1280px) and (min-width: 1080px) {
-  body.single-post .entry-content {
-    --block-wide-max-width: 72vw;
-  }
-}
-
-@media screen and (max-width: 1080px) and (min-width: 960px) {
-  body.single-post .entry-content {
-    --block-wide-max-width: 700px;
-  }
-}
-
-.post {
-  --container-width-base: calc(90vw - var(--frame-size, 0px) * 2);
-  --container-width: var(--container-width-base);
-  max-width: 1440px;
-  width: var(--container-width);
-  font-size: var(--rpjr-text-scale);
-  margin-left: auto;
-  margin-right: auto;
-  padding: 40px 0px;
-}
-
-.icon-rpjr-icon-tag .path1:before {
-  content: '\e907';
-  color: var(--theme-color) !important;
-}
-
-.icon-rpjr-icon-tag .path2:before {
-  content: '\e908';
-  margin-left: -1em;
-  color: rgb(255, 255, 255);
-}
-
-.post .topic-breadcrumb a {
-  width: fit-content;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  border: 3px solid rgb(234, 234, 234);
-  border-radius: 12px;
-  padding: 5px 14px 8px 14px;
-  font-size: 16px;
-  font-weight: 700;
-}
-
-.post .topic-breadcrumb img {
-  margin-right: 15px;
-}
-
-.post .topic-breadcrumb {
-  max-width: 800px !important;
-  margin-right: auto;
-  margin-left: auto;
-}
-
-.post .hero-section {
-  max-width: 800px !important;
-  margin-right: auto;
-  margin-left: auto;
-  margin-bottom: var(--margin-bottom, 40px);
-}
-
-@media screen and (max-width: 689px) {
-  .post .ct-featured-image {
-    max-width: none !important;
-    width: 100% !important;
-  }
-  .post .ct-featured-image figcaption {
-    padding: 0 4vw;
-  }
-}
-
-@media screen and (min-width: 690px) and (max-width: 991px) {
-  .post .ct-featured-image {
-    position: relative;
-    max-width: none !important;
-    width: 120vw !important;
-    left: calc(-35px - 10vw);
-  }
-  .post .ct-featured-image figcaption {
-    padding: 0 calc(75px + 10vw) 0 10vw;
-  }
-}
-
-@media screen and (min-width: 992px) {
-  .post .ct-featured-image {
-    max-width: 1000px !important;
-    margin-right: auto;
-    margin-left: calc(max(90px, 9vw));
-  }
-}
-
-.post .post-intro {
-  max-width: 760px !important;
-  margin-right: auto;
-  margin-left: auto;
-}
-
-.post .title {
-  --lineHeight: 1.3;
-  --heading-color: var(--paletteColor4);
-  font-family: SweiMarkerSansCJKtc-Regular, 'Noto Sans TC', Sans-Serif, serif;
-  font-size: 32px;
-  color: var(--heading-color);
-  line-height: var(--lineHeight);
-  letter-spacing: var(--letterSpacing);
-}
-
-@media screen and (min-width: 768px) {
-  .post .title {
-    font-size: 32px;
-  }
-}
-
-@media screen and (min-width: 1440px) {
-  .post .title {
-    font-size: 36px;
-  }
-}
-
-.post .title.large {
-  font-size: 34px;
-}
-
-@media screen and (min-width: 768px) {
-  .post .title.large {
-    font-size: 34px;
-  }
-}
-
-@media screen and (min-width: 1440px) {
-  .post .title.large {
-    font-size: 38px;
-  }
-}
-
-.post .subtitle {
-  color: var(--theme-color);
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 160%;
-  letter-spacing: 0.08em;
-}
-
-@media screen and (min-width: 768px) {
-  .post .subtitle {
-    font-size: 22px;
-  }
-}
-
-@media screen and (min-width: 1440px) {
-  .post .subtitle {
-    font-size: 24px;
-  }
-}
-
-.post .post-date {
-  font-weight: 500;
-  font-size: 16px;
-  line-height: 160%;
-  letter-spacing: 0.05em;
-  color: #a3a3a3;
-}
-
-.post .post-date-category {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  height: 34px;
-  margin-top: 60px !important;
-}
-
-.post .post-date-category .post-date {
-  font-weight: 500;
-  font-size: 16px;
-  line-height: 160%;
-  letter-spacing: 0.05em;
-  color: #a3a3a3;
-}
-
-.post .post-intro {
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 200%;
-  letter-spacing: 0.05em;
-  color: #575757;
-  padding: 13px 15px;
-  background: #f4f4f4;
-}
-
-@media screen and (min-width: 768px) {
-  .post .post-intro {
-    font-size: 18px;
-    line-height: 200%;
-    padding: 15px 20px;
-    border-radius: 30px;
-  }
-}
-
-@media screen and (min-width: 992px) {
-  .post .post-intro {
-    font-size: 20px;
-    line-height: 200%;
-    padding: 20px 30px;
-  }
-}
-
-.post .post-intro p {
-  margin-bottom: var(--content-spacing);
-}
-
-.post .rpjr-post-tags {
-  margin-top: 80px;
-  margin-bottom: 80px;
-}
-
-@media screen and (min-width: 768px) {
-  .post .rpjr-post-tags {
-    margin-top: 100px;
-    margin-bottom: 100px;
-  }
-}
-
-@media screen and (min-width: 992px) {
-  .post .rpjr-post-tags {
-    margin-top: 60px;
-    margin-bottom: 60px;
-  }
-}
-
-.post .rpjr-post-tags__box {
-  margin: 25px auto 0;
-  display: flex;
-  width: 100%;
-  max-width: 560px;
-  flex-wrap: wrap;
-  column-gap: 15px;
-  row-gap: 30px;
-  align-items: center;
-  justify-content: center;
-}
-
-@media screen and (max-width: 767px) {
-  .post .rpjr-post-tags__box {
-    padding: 0 15px;
-  }
-}
-
-.post .rpjr-post-tags__heading {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-weight: 700;
-  font-size: 18px;
-  line-height: 30px;
-  letter-spacing: 0.08em;
-  color: #232323;
-  vertical-align: middle;
-}
-
-.post .rpjr-post-tags__heading i {
-  letter-spacing: 0 !important;
-}
-
-.post .rpjr-post-tags__heading > i {
-  font-size: 24px;
-}
-
-@media screen and (max-width: 1000px) {
-  .post .social-sidebar-container {
-    display: none;
-  }
-}
-
-@media screen and (min-width: 1001px) {
-  .post .mobile-social-sidebar-container {
-    display: none;
-  }
-}
-
-.post .social-sidebar-container {
-  position: -webkit-sticky;
-  position: sticky;
-  width: fit-content;
-  top: 180px;
-  left: 0;
-  margin-top: -430px;
-}
-
-.post .social-sidebar-container > * {
-  position: relative;
-  left: 0;
-}
-
-.post .mobile-social-sidebar-container {
-  position: fixed;
-  width: 100%;
-  margin: 0 auto;
-  left: 0;
-  bottom: 10px;
-  z-index: 10;
-}
-
-@media screen and (max-width: 767px) {
-  .post
-    .author-section:not(.author-section--show)
-    .author-section__box
-    > *:nth-child(n + 2) {
-    display: none;
-  }
-}
-
-.post .author-section.author-section--show .author-section__m-btn {
-  display: none !important;
-}
-
-.post .author-section__title {
-  font-family: SweiMarkerSansCJKtc-Regular, 'Noto Sans TC', Sans-Serif, serif;
-  display: block;
-  font-weight: 700;
-  font-size: 20px;
-  line-height: 160%;
-  margin: 80px auto 50px;
-  text-align: center;
-}
-
-@media screen and (min-width: 768px) {
-  .post .author-section__title {
-    font-size: 24px;
-  }
-}
-
-@media screen and (min-width: 1440px) {
-  .post .author-section__title {
-    font-size: 28px;
-  }
-}
-
-.post .author-section__box {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: space-evenly;
-  gap: 40px;
-  flex-basis: 33%;
-  max-width: 1100px;
-  width: 100%;
-  margin: 0 auto;
-}
-
-.post .author-section__m-btn {
-  display: block;
-  margin: 30px auto 60px;
-  padding: 8px 12px;
-  cursor: pointer;
-}
-
-@media screen and (min-width: 768px) {
-  .post .author-section__m-btn {
-    display: none;
-  }
-}
-
-body.single-post
-  main
-  article.post
-  > .entry-content
-  > *:not(.alignwide):not(.alignfull) {
-  max-width: 700px;
-}
