+++ packages/frontend/src/modules/article/utils/parse-post-to-content.ts
@@ -1,0 +1,154 @@
+import { GetPostQuery } from '__generated__/operations/post.generated'
+
+import { Author } from '@/components/author-card'
+import { PostSummary } from '@/components/types'
+import {
+  AUTHOR_ROLES_IN_ORDER,
+  AuthorRole,
+  DEFAULT_AVATAR,
+  DEFAULT_THEME_COLOR,
+  Theme,
+} from '@/constants'
+import { getPostSummaries } from '@/utils'
+
+import { AuthorGroup } from '../brief'
+
+function parsePostToContent(post: NonNullable<GetPostQuery['post']>) {
+  // Assemble authors for brief
+  const authorsJSON = post?.authorsJSON as {
+    id: string
+    name: string
+    role: string
+    type: 'link' | 'string'
+  }[]
+  const authorsInBrief: AuthorGroup[] = []
+  let currentAuthorRole = '',
+    currentAuthors: { name: string; link: string }[] = []
+
+  authorsJSON?.forEach((authorJSON, index) => {
+    const author = post?.authors?.find((a) => a?.id === authorJSON?.id)
+    const authName = author ? (author.name ?? '') : authorJSON.name
+    const authorObj = author
+      ? {
+          name: authName,
+          link: `/author/${author.slug}`,
+        }
+      : {
+          name: authName,
+          link: '',
+        }
+    if (index === 0 || authorJSON.role === authorsJSON[index - 1]?.role) {
+      currentAuthorRole = authorJSON.role
+      currentAuthors.push(authorObj)
+    } else {
+      authorsInBrief.push({ title: currentAuthorRole, authors: currentAuthors })
+      currentAuthorRole = authorJSON.role
+      currentAuthors = [authorObj]
+    }
+
+    if (index === authorsJSON?.length - 1) {
+      authorsInBrief.push({ title: currentAuthorRole, authors: currentAuthors })
+    }
+  })
+
+  // Assemble ordered authors for AuthorCard
+  type AuthorWithLink = Author & { link: string }
+  const authors: AuthorWithLink[] =
+    post?.authors
+      ?.map((author) => {
+        const authorJSON = authorsJSON.find(
+          (authorJSON) => authorJSON.id === author?.id
+        )
+        if (!authorJSON) {
+          return undefined
+        }
+        const avatarURL = author?.avatar?.resized?.tiny
+        return {
+          slug: author.slug,
+          name: author.name ?? '',
+          avatar: avatarURL ?? DEFAULT_AVATAR,
+          bio: author.bio ?? '',
+          role: authorJSON.role as AuthorRole,
+          link: authorJSON.type === 'link' ? `/author/${author.slug}` : '',
+        }
+      })
+      .filter((author) => author !== undefined) ?? []
+
+  // Sort authors by AUTHOR_ROLES_IN_ORDER
+  const orderedAuthors = authors
+    ?.filter((author: AuthorWithLink) => author?.link)
+    ?.map((author: AuthorWithLink) => {
+      const roles = author?.role?.split('、')
+      const priority = AUTHOR_ROLES_IN_ORDER.indexOf(roles?.[0] as AuthorRole)
+      return {
+        ...author,
+        priority: priority === -1 ? AUTHOR_ROLES_IN_ORDER.length : priority,
+      }
+    })
+    ?.sort((a, b) => {
+      return a.priority - b.priority
+    })
+
+  // Topic related data
+  const topic = post?.projects?.[0]
+
+  // Related posts data: related posts or topic's related post
+  let relatedPosts: ReturnType<typeof getPostSummaries> = []
+  if (
+    post?.relatedPostsOrdered?.length &&
+    post?.relatedPostsOrdered?.length > 0
+  ) {
+    relatedPosts = getPostSummaries(post.relatedPostsOrdered ?? [])
+  } else if (topic?.relatedPosts?.length && topic?.relatedPosts?.length > 0) {
+    relatedPosts = getPostSummaries(topic?.relatedPosts ?? [])
+  }
+
+  // Main project data
+  // TODO: project/main project are duplicate data, should be refactored
+  const mainTopic = post?.mainProject
+  const topicURL = mainTopic?.slug ? `/topic/${mainTopic.slug}` : undefined
+
+  // Subcategory related data
+  const subSubcategory = post?.subSubcategoriesOrdered?.[0]
+  const subcategory = subSubcategory?.subcategory
+  const category = subcategory?.category
+  const subSubcategoryURL =
+    category?.slug && subcategory?.slug && subSubcategory?.slug
+      ? `/category/${category.slug}/${subcategory.slug}/${subSubcategory.slug}`
+      : ''
+  const theme = (category?.themeColor || DEFAULT_THEME_COLOR) as Theme
+
+  const twReporterRelatedPosts: PostSummary[] =
+    post?.TWReporterRelatedPostsJSON?.map(
+      (twReporterPost: {
+        ogTitle: string
+        src: string
+        ogImgSrc: string
+        ogDescription: string
+        publishedDate: string
+      }) => ({
+        title: twReporterPost.ogTitle,
+        url: twReporterPost.src,
+        image: twReporterPost.ogImgSrc,
+        desc: twReporterPost.ogDescription,
+        category: '',
+        subSubcategory: '',
+        publishedDate: twReporterPost.publishedDate,
+        theme: DEFAULT_THEME_COLOR,
+      })
+    ) ?? []
+
+  return {
+    theme,
+    topicURL,
+    mainTopic,
+    subSubcategory,
+    subSubcategoryURL,
+    authorsInBrief,
+    orderedAuthors,
+    relatedPosts,
+    twReporterRelatedPosts,
+  }
+}
+
+export default parsePostToContent
